+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   - `#198`: As scp user, I want to opt for a prompt if `username` is missing. (`option.prompt`)
   - `#201`: As scp user, I want to opt for a prompt if `password` is missing.
   - `#204`: As scp user, I want to opt for a prompt if `host` is missing.
+- **Test**
+  - `#208`: As test, I want to refactor `username` tests to use `mockResolvedValueOnce`
 
 ## `0.13.1` - 2021-Aug-4